pyproject.toml
[tool.poetry]
name = "deadonfilm"
version = "0.1.0"
description = ""
authors = ["<chris@waitingforthefuture.org>"]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.1"
dateparser = "0.7.0"
requests = "^2.26.0"
tqdm = "^4.62.2"
fastapi = {extras = ["all"], version = "^0.101.1"}
black = "^23.7.0"
sqlalchemy = "^2.0.20"
cinemagoer = "^2023.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
